+++ CHANGELOG.md
@@ -12,6 +12,7 @@ PDBTools.jl Changelog
 
 Version 3.32.0-DEV
 --------------
+- ![FEATURE][badge-feature] Allow computation of `mvalue`s and `transfer_free_energy`s from precomputed SASA values. For that, dispatch on the `CreamerUnitedAtomRadii` type the SASA calculation, for consistency. Otherwise, an error is thrown.  
 - ![FEATURE][badge-feature] Support for `unitcell` keyword argument in `transfer_free_energy` and `mvalue` functions, which will compute SASAs with periodic boundary conditions. 
 
 Version 3.31.2